Fix typos, add 'icon' and 'precision' updates.
-Fix a lot of typos.

-Adds 'icon:"mdi:current-dc"' to struct fields representing current
values in various BMS and inverter protocols for better
visualization in Home Assistant.

-Adds or updates the `precision` attribute in Go struct tags for fields
that have a `multiplier` attribute. The `precision` value is set to
the number of decimal places in the `multiplier`. This is needed in
Homeassistant since July. Setting the precision will output a
'suggested_display_precision' attribute in the discovery information
publiched to MQTT. HA might need to be refreshed, MQTT restarted for
discovery to realize the change.